[x86_64] Abstract some architecture-specific stuff away from assembly emission
