+++ CMakeLists.txt
@@ -67,6 +67,7 @@ add_executable(
   src/codegen/register_allocation.c
   src/codegen/ir/ir.c
   src/codegen/x86_64/arch_x86_64.c
+  src/codegen/x86_64/arch_x86_64_common.c
 )
 target_include_directories(
   intc
